# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "KKN Integration Dashboard API",
        "description": "Dokumentasi API dashboard KKN. API ini dipakai untuk integrasi dengan aplikasi project-akhir (portal mahasiswa & dosen) yang mengirim data pendaftaran, dokumen, dan nilai ke dashboard.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8001",
            "description": "Server lokal dashboard"
        }
    ],
    "paths": {
...
        "/api/form-kesediaan/{id}": {
            "get": {
                "tags": [
                    "Form Kesediaan"
                ],
                "summary": "Detail form kesediaan",
                "description": "Mengambil detail satu form kesediaan berdasarkan ID.",
                "operationId": "f7d2894746e008ede4aaa35e3d8749c3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID form kesediaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail form kesediaan berhasil diambil"
                    },
                    "404": {
                        "description": "Data tidak ditemukan"
                    }
                }
            }
        },
        "/api/form-kesediaan/{id}/status": {
            "put": {
                "tags": [
                    "Form Kesediaan"
                ],
                "summary": "Verifikasi / ubah status form kesediaan",
                "description": "Mengubah status verifikasi form kesediaan (pending, approved, rejected) beserta catatan opsional.",
                "operationId": "2e47497ed6055bf04f09f2499a2f049e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID form kesediaan",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "status"
                                ],
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "approved",
                                        "enum": [
                                            "pending",
                                            "approved",
                                            "rejected"
                                        ]
                                    },
                                    "catatan": {
                                        "type": "string",
                                        "example": "Disetujui",
                                        "nullable": true,
                                        "maxLength": 1000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status form kesediaan berhasil diperbarui"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                }
            }
        },
        "/api/form-kesediaan/store-from-external": {
            "post": {
                "tags": [
                    "Form Kesediaan"
                ],
                "summary": "Menerima form kesediaan dari project-akhir",
                "description": "Menyimpan form kesediaan dosen beserta file (dikirim sebagai string base64). Endpoint ini juga tersedia sebagai alias di /api/form-kesediaan/store.",
                "operationId": "824d7a7f2a82750d9994fd9be67f8cc6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "judul_kegiatan",
                                    "user_nim",
                                    "file_content",
                                    "file_name",
                                    "file_mime_type",
                                    "file_size"
                                ],
                                "properties": {
                                    "judul_kegiatan": {
                                        "type": "string",
                                        "example": "KKN Tematik Desa Sukamaju",
                                        "maxLength": 255
                                    },
                                    "user_nim": {
                                        "type": "string",
                                        "example": "2010001",
                                        "maxLength": 20
                                    },
                                    "file_content": {
                                        "description": "Isi file PDF dalam base64",
                                        "type": "string",
                                        "format": "byte"
                                    },
                                    "file_name": {
                                        "type": "string",
                                        "example": "form-kesediaan.pdf",
                                        "maxLength": 255
                                    },
                                    "file_mime_type": {
                                        "type": "string",
                                        "example": "application/pdf"
                                    },
                                    "file_size": {
                                        "type": "integer",
                                        "example": 204800
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Form kesediaan berhasil disimpan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "500": {
                        "description": "Terjadi kesalahan server"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Masukkan API key mitra di sini untuk endpoint tarik data nilai.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}